Move tracing exporter registration
This doesn't belong in the library and should be configured by the
consumer of the opencensus package.
