+++ cmd/virtual-kubelet/commands/root/tracing_register.go
@@ -1,4 +1,4 @@
-package opencensus
+package root
 
 import (
 	"github.com/virtual-kubelet/virtual-kubelet/errdefs"
