+++ cmd/virtual-kubelet/commands/root/tracing_register_ocagent.go
@@ -1,6 +1,6 @@
 // +build !no_ocagent_exporter
 
-package opencensus
+package root
 
 import (
 	"os"
