+++ cmd/virtual-kubelet/commands/root/tracing_register_test.go
@@ -1,4 +1,4 @@
-package opencensus
+package root
 
 import (
 	"testing"
